+++ NEWS.md
@@ -10,7 +10,9 @@
 * added more datasets (files) to test functionality
 * building a functionality to create separate models for each asset (adding functions with prefix aml_xxx_xxx.R)
 * build more simple functions to collect data, build model, score data to predict
-* added tests for functions collect data and test model
+* added tests for several functions
+* function aml_make_model() will be finding the best neural network structure automatically using 6 random structures
+* fixing of warnings due to column name assignments
 
 # lazytrade 0.3.4