+++ NEWS.md
@@ -1,6 +1,13 @@
 # lazytrade (development version)
 
-* removal of duplicate rows in the collect data function
+# lazytrade 0.3.6
+
+# Version 0.3.6
+
+## Changes
+
+* removal of duplicate rows in the `aml_collect_data()` function
+* added a test for `to_m()` function
 
 # lazytrade 0.3.5